Minor changes for consistency
The name of the overlay layer for the bike racks is now passed as a
string, in case more overlays are added in the future, to avoid
inconsistencies. When the markers are added to the layer in the
controller the same string is used, to avoid any possible mishaps in
the naming of the layer. The same should be done for any future layers.